% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{Leveraging}
\usage{
algo_leverage(Y, X, rsize, method)
}
\arguments{
\item{Y}{= response vector}

\item{X}{= covariate data frame}

\item{rsize}{= the subset sample size}

\item{method}{= either "Uniform" or "Leverage" for weighting}
}
\value{
output of coefficient estimates, with t-statistic and p-value
}
\description{
This function estimates regression coefficients using a randomly selected subset of the full dataset.
}
\examples{
set.seed(123)
fullX <- rt(500, df=6)
fullY <- -1*fullX + rnorm(100)
fullDF <- data.frame(fullY, fullX)
plot(fullX, fullY, main="Full Sample")

summary(lm(fullY ~ fullX))$coef #True coefficients, compare with uniform/leverage output

algo_leverage(fullY, data.frame(fullX), rsize = 200, method="Uniform")
algo_leverage(fullY, data.frame(fullX), rsize = 200, method="Leverage")

}
\keyword{leveraging}
