% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{elnet_coord}
\alias{elnet_coord}
\title{Elastic Net Coordinate Descent}
\usage{
elnet_coord(Y, X, alpha, B)
}
\arguments{
\item{Y}{= response vector}

\item{X}{= covariate matrix}

\item{alpha}{= parameter between [0,1]}

\item{B}{=  initial vector}
}
\value{
a matrix with lambda values in the first column, and corresponding coefficient estimates
}
\description{
This function estimates coefficients for elastic net and also has function to visualize solution path.
}
\examples{
library(MASS)
 set.seed(123)
 n <- 100
 mu <- rep(0, 20)
 Sigma <- diag(20)
Sigma[1,2] <- Sigma[2,1] <- Sigma[5,6] <- Sigma[6,5] <- 0.8
TrueBeta <- c(2, 0, -2, 0, 1, 0, -1, 0, rep(0, times=12))
X <- mvrnorm(n, mu, Sigma)
Y <- X\%*\%TrueBeta + rnorm(n)

require(ggplot2)
require(reshape2)
output <- elnet_coord(Y, X, alpha=0, B=rep(0, times=ncol(X)))
visualize_coord(output)
output <- elnet_coord(Y, X, alpha=0.5, B=rep(0, times=ncol(X)))
visualize_coord(output)
output <- elnet_coord(Y, X, alpha=1, B=rep(0, times=ncol(X)))
visualize_coord(output)




}
\keyword{coordinate}
\keyword{descent}
\keyword{elastic}
\keyword{net,}
