\name{hw2_data}
\alias{hw2_data}
\docType{data}
\title{
hw2 data used for testing
}
\description{
hw2 data used for testing mixture function
}
\usage{data(hw2_data)}
\format{
  A data frame with 19599 observations on the following variable.
  \describe{
    \item{\code{X17.5439121669982}}{a numeric vector}
  }
}

\keyword{datasets}
