% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Fit Ridge Regression}
\usage{
ridge_regression(form, data, lambda = 0)
}
\arguments{
\item{form}{The formula for linear model. Example: y ~ x.}

\item{data}{The input dataset.}

\item{lambda}{regulation parameter, the larger the lambda the more penalization on the coefficients, default = 0}
}
\value{
The result model which contains the coefficients.
}
\description{
Fit a ridge regression model based on the provided formula, data and lambda.
}
\examples{
fit_ridge <- ridge_regression(Sepal.Length ~ ., iris)
fit_ridge <- ridge_regression(Sepal.Length ~ ., iris, lambda = 0.01)
}
\author{
Michael Kane & Tong Wei
}
