% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ridge_regression_lambda_selection}
\alias{ridge_regression_lambda_selection}
\title{Perform ridge regression and select the optimal lambda}
\usage{
ridge_regression_lambda_selection(form, data, lambda = seq(0, nrow(data)
  * 2, length.out = 500))
}
\arguments{
\item{form}{A formilar.}

\item{data}{A dataset.}

\item{lambda}{A tuning parameter.}
}
\value{
Construct a ridge regression model containing calculated coefficients and use mse to find the optimal lambda.
}
\description{
Perform ridge regression and select the optimal lambda
}
\examples{
ridge_regression_lambda_selection(Sepal.Length~., iris)
}
