% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opimizing.R
\name{ridge_regression_optlambda}
\alias{ridge_regression_optlambda}
\title{Fit Ridge Regression with optimizing lambda}
\usage{
ridge_regression_optlambda(form, data, lambda_seq = seq(0, 200 * 2,
  length.out = 500))
}
\arguments{
\item{form}{The formula for linear model. Example: y ~ x.}

\item{data}{The input dataset.}

\item{lambda_seq}{the sequence of the tested regulation parameter, default = seq(0, 200*2, length.out = 500)}
}
\value{
The result model which contains the coefficients.
}
\description{
Fit a ridge regression model based on the provided formula, data.
}
\examples{
fit_ridge <- ridge_regression(Sepal.Length ~ ., iris)
}
\author{
Tong Wei
}
