% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxi}
\alias{taxi}
\title{testing data}
\format{A data frame with 10000 rows and 14 variables:
\describe{
  \item{obs_id}{the taxi id}
  \item{train_id}{whether is in train or test}
  \item{duration}{duration of the trip}
  \item{day}{days in number}
  \item{hour}{hours in number}
  \item{minute}{minute in number}
  \item{weekday}{weekdays in number}
  \item{pickup_longitude}{pickup_longitude}
  \item{pickup_latitude}{pickup_latitude}
  \item{dropoff_longitude}{dropoff_longitude}
  \item{dropoff_latitude}{dropoff_latitude}
  \item{trip_distance}{trip_distance}
  \item{pickup_NTACode}{pickup_NTACode}
  \item{dropoff_NTACode}{dropoff_NTACode}
}}
\usage{
taxi
}
\description{
The taxi dataset the to test the ridge regression function
}
\keyword{datasets}
