% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_solve.R
\name{tridiagonal}
\alias{tridiagonal}
\title{Tridiagonal matrix}
\usage{
tridiagonal(d, offd, n)
}
\arguments{
\item{d}{the value on the diagonal}

\item{offd}{the value along the off diagonal}

\item{n}{the dimension of the matrix}
}
\value{
A, an n x n tridiagonal matrix
}
\description{
Generates a symmetric tridiagonal matrix with "offd" on off diagonal, "d" on the diagonal
}
