% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_coord.R
\name{visualize_coord}
\alias{visualize_coord}
\title{Visualizing Elastic Net}
\usage{
visualize_coord(solpath)
}
\arguments{
\item{solpath}{= the output of elnet_coord()}
}
\value{
a plot of the solution path similar to enet package
}
\description{
This function visualizes the solution path, given by the output of elnet_coord()
}
\examples{
require(ggplot2)
require(reshape2)
output <- elnet_coord(Y, X, alpha=0, B=rep(0, times=ncol(X)))
visualize_coord(output)
output <- elnet_coord(Y, X, alpha=0.5, B=rep(0, times=ncol(X)))
visualize_coord(output)
output <- elnet_coord(Y, X, alpha=1, B=rep(0, times=ncol(X)))
visualize_coord(output)
}
\keyword{path}
\keyword{solution}
\keyword{visualizing}
