\name{Homework2-package}
\alias{Homework2-package}
\alias{Homework2}
\docType{package}
\title{
MLE and stderr for mixture model of two normal components
}
\description{
This package contains one function "mixture" that calculates MLE and standard error for mixture model parameters using either EM or newton method.
}
\details{
\tabular{ll}{
Package: \tab Homework2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-20\cr
License: \tab GPL-2\cr
}
}
\author{
Yuxin(Daisy) Zhu

Maintainer: Yuxin(Daisy) Zhu <yuzhu@jhsph.edu>
}
\keyword{ mixture normal}
\keyword{  EM}% __ONLY ONE__ keyword per line
\keyword{newton}