\name{Homework2-package}
\alias{Homework2-package}
\alias{Homework2}
\docType{package}
\title{ML estimation using NR and EM algorithms.}
\description{Return MLEs for a mixture distribution with two normal components}
\details{
\tabular{ll}{
Package: \tab Homework2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-04\cr
License: \tab GPL-2\cr
}

}
\author{Oliver (slightly) Awesome
}
