\name{Homework2-package}
\alias{Homework2-package}
\alias{Homework2}
\docType{package}
\title{
Homework 2
}
\description{
Advanced Statistics Computing Homework 2 which includes EM algorithm and Newton algorithm.
}
\details{
\tabular{ll}{
Package: \tab Homework2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-03\cr
License: \tab GPL\cr
}
run mixture() function to estimate parameters for two mixed normal model.

}
\author{
Yu Du

Maintainer: Yu Du<ydu@jhsph.edu>

}
\references{
Advanced Statistics Computing Class. Dr. Peng.
}
\keyword{ package }
