% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Perform ridge regression}
\usage{
ridge_regression(form, data, lambda = 0)
}
\arguments{
\item{form}{A formilar.}

\item{data}{A dataset.}

\item{lambda}{A tuning parameter.}
}
\value{
Construct a ridge regression model containing calculated coefficients.
}
\description{
Perform ridge regression
}
\examples{
ridge_regression(Sepal.Length~., iris)
}
